example updated
