+++ example/android/app/build.gradle
@@ -123,7 +123,7 @@ dependencies {
     compile project(':react-native-sensitive-info')
     compile fileTree(dir: "libs", include: ["*.jar"])
     compile "com.android.support:appcompat-v7:23.4.0"
-    compile "com.facebook.react:react-native:0.20.1"  // From node_modules
+    compile "com.facebook.react:react-native:0.27.2"  // From node_modules
 }
 
 // Run this once to be able to run the application with BUCK
